@@ -28,12 +28,13 @@
     "@codemirror/lang-javascript": "^6.0.1",
     "@codemirror/lang-python": "^6.0.0",
     "@codemirror/theme-one-dark": "^6.0.0",
-    "@headlessui/react": "^1.7.3",
+    "@headlessui/react": "^1.7.11",
     "@popperjs/core": "^2.11.2",
-    "@reduxjs/toolkit": "^1.7.2",
-    "@types/react-dom": "^17.0.11",
+    "@redux-devtools/core": "^3.13.3",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@types/react-dom": "^18.2.15",
     "@types/ua-parser-js": "^0.7.36",
-    "@webscopeio/react-textarea-autocomplete": "^4.7.3",
+    "@webscopeio/react-textarea-autocomplete": "^4.9.2",
     "ace-builds": "^1.4.14",
     "acorn": "^8.10.0",
     "acorn-walk": "^7.2.0",
@@ -44,8 +45,8 @@
     "d3": "^7.6.1",
     "dayjs": "^1.10.7",
     "fastest-levenshtein": "^1.0.16",
-    "html-react-parser": "^0.13.0",
-    "i18next": "^20.6.1",
+    "html-react-parser": "^5.0.6",
+    "i18next": "^23.7.6",
     "i18next-browser-languagedetector": "^6.1.3",
     "js-interpreter": "^5.0.0",
     "jszip": "^3.8.0",
@@ -54,19 +55,17 @@
     "lunr": "^2.3.9",
     "prop-types": "^15.8.1",
     "re-reselect": "^4.0.0",
-    "react": "^16.14.0",
+    "react": "^18.2.0",
     "react-confetti": "^6.1.0",
-    "react-dom": "^16.14.0",
-    "react-hot-loader": "^4.13.0",
-    "react-i18next": "^11.15.4",
-    "react-popper": "^2.2.5",
-    "react-redux": "^7.2.6",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^13.4.1",
+    "react-popper": "^2.3.0",
+    "react-redux": "^8.1.3",
     "react-split": "^2.0.14",
     "react-use": "^17.4.0",
-    "react-virtualized-auto-sizer": "^1.0.6",
-    "react-window": "^1.8.6",
+    "react-virtualized-auto-sizer": "^1.0.20",
+    "react-window": "^1.8.9",
     "redux": "^4.1.2",
-    "redux-devtools": "^3.7.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
@@ -86,31 +85,32 @@
     "@tailwindcss/forms": "^0.5.0",
     "@tailwindcss/typography": "^0.5.2",
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^14.1.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/classnames": "^2.3.1",
     "@types/gtag.js": "0.0.10",
     "@types/jest": "^27.0.2",
     "@types/lodash": "^4.14.178",
     "@types/lunr": "^2.3.4",
-    "@types/react": "^17.0.39",
-    "@types/react-redux": "^7.1.22",
-    "@types/react-virtualized-auto-sizer": "^1.0.1",
-    "@types/react-window": "^1.8.5",
+    "@types/react": "^18.2.37",
+    "@types/react-redux": "^7.1.30",
+    "@types/react-virtualized-auto-sizer": "^1.0.3",
+    "@types/react-window": "^1.8.8",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
     "autoprefixer": "^10.4.2",
+    "axe-core": "^4.8.2",
     "babel-jest": "^27.3.1",
     "babel-loader": "^8.2.3",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^6.6.0",
-    "cypress": "^11.2.0",
+    "cypress": "^13.6.3",
     "cypress-axe": "^1.5.0",
     "cypress-file-upload": "^5.0.8",
     "cypress-mochawesome-reporter": "^3.1.0",
     "cypress-multi-reporters": "^1.6.0",
-    "cypress-network-idle": "^1.10.2",
-    "cypress-real-events": "^1.7.4",
+    "cypress-network-idle": "^1.14.2",
+    "cypress-real-events": "^1.11.0",
     "dotenv-webpack": "^5.1.0",
     "eslint": "^8.9.0",
     "eslint-config-standard": "^17.1.0",
