@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react"
-import { useDispatch, useSelector } from "react-redux"
+import { useAppDispatch as useDispatch, useAppSelector as useSelector } from "../hooks"
 import { Collapsed } from "../browser/Utils"
 
 import ReactTextareaAutocomplete from "@webscopeio/react-textarea-autocomplete"
