@@ -17,7 +17,7 @@ describe("Editor", () => {
         cy.get('[data-test="newScript"]').click()
         cy.get("#scriptName").type("cypress_test")
         cy.get("input").contains("CREATE").click()
-        cy.get("div[id^='headlessui-dialog-']", { timeout: 10000 }).should("not.exist")
+        cy.contains("div", "Create a new script", { timeout: 10000 }).should("not.exist")
     })
 
     it("runs template script", () => {
